backend/requirements.txt
Django
djangorestframework
django-rest-auth
django-cors-headers
psycopg2-binary
pandas
scikit-learn==0.23.2
joblib
gunicorn
Cpython